+++ include/aikido/robot/Robot.hpp
@@ -308,8 +308,10 @@ public:
   /// Releases DoFs held by executor if held.
   ///
   /// \param[in] executor The Executor to add to the inactive list.
-  /// \return A robot-unique non-negative ID (negative implies failure)
-  virtual int registerExecutor(aikido::control::ExecutorPtr executor);
+  /// \param[in] desiredName The desired name for the executor.
+  /// \return A robot-unique non-negative ID (negative implies failure).
+  virtual int registerExecutor(
+      aikido::control::ExecutorPtr executor, std::string desiredName = "");
 
   ///
   /// Deactivates the current active executor.
@@ -319,7 +321,19 @@ public:
   ///
   /// \param[in] id of executor on executor list
   /// \return True if successful. If false, all executors are inactive.
-  virtual bool activateExecutor(int id);
+  virtual bool activateExecutor(const int id);
+
+  ///
+  /// Convenience:
+  /// Activates executor using name
+  /// Deactivates the current active executor.
+  /// Sets an executor from the inactive executor list to be the
+  /// active executor.
+  /// Holds and releases DoFs as needed.
+  ///
+  /// \param[in] id of executor on executor list
+  /// \return True if successful. If false, all executors are inactive.
+  virtual bool activateExecutor(const std::string name);
 
   ///
   /// Convenience:
@@ -423,8 +437,13 @@ protected:
   std::string mName;
   dart::dynamics::MetaSkeletonPtr mMetaSkeleton;
   aikido::statespace::dart::MetaSkeletonStateSpacePtr mStateSpace;
+
+  // Currently active executor
   int mActiveExecutor{-1};
+  // Executors indexed by id (higher id has more priority)
   std::vector<aikido::control::ExecutorPtr> mExecutors;
+  // Keeps track of executor name (if used)
+  std::unordered_map<std::string, int> mExecutorsNameMap;
 
   // Subrobot and Joint Management
   Robot* mParentRobot{nullptr};
@@ -453,6 +472,18 @@ protected:
   using ConfigurationMap
       = std::unordered_map<std::string, const Eigen::VectorXd>;
   ConfigurationMap mNamedConfigurations;
+
+  ///
+  /// Checks validity of subrobot for registration:
+  /// name should be unique.
+  /// All body nodes in skeleton should be owned by this robot.
+  /// Subrobot DoFs should be disjoint.
+  ///
+  /// \param[in] metaSkeleton The referential skeleton corresponding to the
+  /// subrobot. \param[in] name Name of the subrobot.
+  virtual bool validateSubRobot(
+      dart::dynamics::ReferentialSkeletonPtr refSkeleton,
+      const std::string& name);
 
   // Parameters for Default VF Planners
   util::VectorFieldPlannerParameters mVFParams;